+++ src/main/java/dt/repository/DoctorRepository.java
@@ -3,7 +3,7 @@ package dt.repository;
 import dt.model.Doctor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DoctorRepository extends CrudRepository<Doctor,Integer> {
+public interface DoctorRepository extends CrudRepository<Doctor,Long> {
 
-    Doctor findByUserAccountId(Integer userId);
+    Doctor findByUserAccountId(Long userId);
 }
