+++ src/main/java/dt/repository/UserRepository.java
@@ -3,6 +3,6 @@ package dt.repository;
 import dt.model.UserAccount;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserAccount,Integer> {
+public interface UserRepository extends CrudRepository<UserAccount,Long> {
     UserAccount findByEmail(String email);
 }
